Refactor webui into modular architecture and add model lifecycle management
- Refactor monolithic webui/app.py (~2.4k LOC) into modular router packages
- Add collection metadata storage to track model/quantization used for each collection
- Implement lazy model loading with automatic unloading after inactivity
- Fix quantization display in job monitoring UI

Breaking changes:
- None, backward compatibility maintained

New features:
- Model lifecycle management with configurable inactivity timeout
- Collection metadata ensures correct model is used for searches
- /model/status endpoint for monitoring model state
