+++ start_webui.sh
@@ -20,4 +20,4 @@ echo "Access the interface at: http://localhost:8080"
 echo ""
 
 # Run the web UI with SQLite-based job tracking
-python3 -m uvicorn webui.app:app --host 0.0.0.0 --port 8080
+poetry run uvicorn webui.app:app --host 0.0.0.0 --port 8080
