+++ webui/api/__init__.py
@@ -1,0 +1,7 @@
+"""
+API router package for Document Embedding Web UI
+"""
+
+from . import auth, jobs, files, metrics, root, settings, models, search
+
+__all__ = ["auth", "jobs", "files", "metrics", "root", "settings", "models", "search"]
