+++ webui/schemas.py
@@ -1,0 +1,13 @@
+"""
+Shared Pydantic models for the Web UI
+"""
+
+from typing import Optional
+from pydantic import BaseModel
+
+class FileInfo(BaseModel):
+    path: str
+    size: int
+    modified: str
+    extension: str
+    hash: Optional[str] = None
